refactor: 장바구니 관리 훅 추가 및 App.tsx에서 장바구니 관련 로직 통합
- useCart 훅을 추가하여 장바구니 상태 관리 및 관련 로직을 분리
- App.tsx에서 기존 장바구니 관련 상태 및 로직을 useCart로 통합
- 불필요한 주석 처리된 코드 삭제 및 코드 정리
