+++ src/basic/App.tsx
@@ -1,11 +1,10 @@
 import { useState, useCallback, useEffect } from 'react';
-import { CartItem, Coupon, Notification, Product, ProductWithUI } from '../types';
+import { CartItem, ProductWithUI } from '../types';
 import Button from './components/ui/Button';
 import IconButton from './components/ui/IconButton';
 import Tab from './components/ui/Tab';
 import Toast from './components/ui/Toast';
 import Header from './components/ui/Header';
-import { initialCoupons } from './data/mockCoupons';
 import { useProducts } from './hooks/product/useProducts';
 import { useNotifications } from './hooks/notifications/useNotifications';
 import { useProductForm } from './hooks/product/useProductForm';
@@ -14,6 +13,7 @@ import { formatPriceWithSoldOut } from './utils/formatters/priceFormatter';
 import { useCoupons } from './hooks/coupons/useCoupons';
 import { calculateCartTotal, calculateItemTotal } from './utils/calculations/cartCalculations';
 import { useCouponsForm } from './hooks/coupons/useCouponsForm';
+import { useCart } from './hooks/cart/useCart';
 
 const App = () => {
   const { products, deleteProduct, updateProduct, addProduct } = useProducts();
@@ -23,29 +23,6 @@ const App = () => {
 
   const { productForm, setProductForm, editingProduct, setEditingProduct, handleProductSubmit } =
     useProductForm(addProduct, updateProduct);
-
-  const [cart, setCart] = useState<CartItem[]>(() => {
-    const saved = localStorage.getItem('cart');
-    if (saved) {
-      try {
-        const parsedCart = JSON.parse(saved);
-        if (Array.isArray(parsedCart)) {
-          return parsedCart.filter(
-            (item) =>
-              item &&
-              item.product &&
-              typeof item.product.id === 'string' &&
-              typeof item.product.price === 'number' &&
-              typeof item.quantity === 'number',
-          );
-        }
-        return [];
-      } catch {
-        return [];
-      }
-    }
-    return [];
-  });
 
   // coupons---------------------!!
   const { coupons, setCoupons, selectedCoupon, setSelectedCoupon, applyCoupon } = useCoupons(
@@ -71,20 +48,12 @@ const App = () => {
 
   // Admin
 
-  const [totalCartItem, setTotalCartItem] = useState(0);
-
-  useEffect(() => {
-    const count = cart.reduce((sum, item) => sum + item.quantity, 0);
-    setTotalCartItem(count);
-  }, [cart]);
-
-  useEffect(() => {
-    if (cart.length > 0) {
-      localStorage.setItem('cart', JSON.stringify(cart));
-    } else {
-      localStorage.removeItem('cart');
-    }
-  }, [cart]);
+  // cart---------------------!!
+  const { cart, setCart, totalCartItem, addToCart, removeFromCart, updateQuantity } = useCart(
+    products,
+    (message) => addNotification(message, 'success'),
+    (message) => addNotification(message, 'error'),
+  );
 
   useEffect(() => {
     const timer = setTimeout(() => {
@@ -92,67 +61,6 @@ const App = () => {
     }, 500);
     return () => clearTimeout(timer);
   }, [searchInputValue]);
-
-  const addToCart = useCallback(
-    (product: ProductWithUI) => {
-      const remainingStock = getRemainingStock(product, cart);
-      if (remainingStock <= 0) {
-        addNotification('재고가 부족합니다!', 'error');
-        return;
-      }
-
-      setCart((prevCart) => {
-        const existingItem = prevCart.find((item) => item.product.id === product.id);
-
-        if (existingItem) {
-          const newQuantity = existingItem.quantity + 1;
-
-          if (newQuantity > product.stock) {
-            addNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
-            return prevCart;
-          }
-
-          return prevCart.map((item) =>
-            item.product.id === product.id ? { ...item, quantity: newQuantity } : item,
-          );
-        }
-
-        return [...prevCart, { product, quantity: 1 }];
-      });
-
-      addNotification('장바구니에 담았습니다', 'success');
-    },
-    [cart, addNotification, getRemainingStock],
-  );
-
-  const removeFromCart = useCallback((productId: string) => {
-    setCart((prevCart) => prevCart.filter((item) => item.product.id !== productId));
-  }, []);
-
-  const updateQuantity = useCallback(
-    (productId: string, newQuantity: number) => {
-      if (newQuantity <= 0) {
-        removeFromCart(productId);
-        return;
-      }
-
-      const product = products.find((p) => p.id === productId);
-      if (!product) return;
-
-      const maxStock = product.stock;
-      if (newQuantity > maxStock) {
-        addNotification(`재고는 ${maxStock}개까지만 있습니다.`, 'error');
-        return;
-      }
-
-      setCart((prevCart) =>
-        prevCart.map((item) =>
-          item.product.id === productId ? { ...item, quantity: newQuantity } : item,
-        ),
-      );
-    },
-    [products, removeFromCart, addNotification, getRemainingStock],
-  );
 
   const completeOrder = useCallback(() => {
     const orderNumber = `ORD-${Date.now()}`;
